Fixed not existing id crashing the app, because no checks were made if an id existed before deleting it

+++ src/main/java/com/bert/test/test/services/DipendentiServiceImpl.java
@@ -48,7 +48,11 @@ public class DipendentiServiceImpl implements DipendentiService{
 
 	@Override
 	public void deleteById(Long id) {
-		dipendentiRepository.deleteById(""+id);
+		if(dipendentiRepository.existsById(""+id)) {
+			dipendentiRepository.deleteById(""+id);
+		}else {
+			return ;
+		}
 	}
 
 	@Override
